% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mulitiwords.R
\name{add_multiwords}
\alias{add_multiwords}
\title{Find multi-word expressions.}
\usage{
add_multiwords(
  word_df,
  tokens,
  min_hits = 3,
  word_field = "words",
  levels = c(1, 2)
)
}
\arguments{
\item{word_df}{A data.frame containing words.}

\item{tokens}{A word-tokens object, returned by quanteda::tokens.}

\item{min_hits}{Numerical. Default is 3. Minimum occurrence of the found multi-word
expressions.}

\item{word_field}{Character. Default is "words". Name of the column in \code{word_df}
that contains the words.}

\item{levels}{Numerical (1 or 2). The window size for multi-word expressions.}
}
\value{
A data.frame with all information from \code{word_df}, but with added multi-word
expressions in rows.
Additionally, the returned data.frame contains the columns...
\itemize{
\item \code{'orig_id'} with a unique identifier for each original word, re-used from \code{word_df} or created new if not present
\item \code{'from'} indicating the word of origin
\item \verb{'word_id} (character) with a unique identifier for all words and multi-words
\item \verb{'hits} indicating the number of occurrences of the word or multi-word
}
}
\description{
Adds multi-word expressions found in a quanteda tokens object to a data.frame of words.
}
\examples{
tw_data \%<>\% head(100) \%>\% clean_text(text_field = 'full_text')
toks <- quanteda::tokens(tw_data$text)
data.frame(words = c("deutschen", "millionen")) \%>\%
            add_multiwords(tokens = toks,
            min_hits = 1,
            levels = 1)
}
