% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_word_id}
\alias{add_word_id}
\title{Add word_id.}
\usage{
add_word_id(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
A data.frame that includes a column called 'word_id' of type 'character' which
contains a unique identifiers for all rows in the data.frame.
}
\description{
Function to add word_id to a data.frame if it is not already there.
}
\details{
Searches the data.frame for variables that include 'id' in their name.
Checks if any of these is a unique identifier. If so, this variable is renamed
'word_id' and forced into type 'character'.
If not, a new column named 'word_id' of type 'character' is created, containing a
string of numbers that uniquely identify all rows in the data.frame.
}
\examples{
t1 <- data.frame(wid = 1:10, a = sample(letters, 10))
t2 <- data.frame(id = rep(1, 10),
                word_id = sprintf("\%02d", 1:10),
                a = sample(letters, 10))
t3 <- data.frame(a = sample(letters, 10))
add_word_id(t1)
add_word_id(t2)
add_word_id(t3)
}
\seealso{
\code{\link{find_unique_id}},  \code{\link{add_multiwords}}
}
