% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Clean text}
\usage{
clean_text(
  df,
  text_field = "text",
  clean_field = "text",
  tolower = T,
  remove_punct = T,
  simplify_punct = F,
  replace_emojis = T,
  replace_numbers = T,
  remove_stopwords = F,
  store_uncleaned = T,
  count = T
)
}
\arguments{
\item{df}{A data frame.}

\item{text_field}{character. Default is 'text'. Name of a column in \code{df} that
contains text that should be cleaned. Default is 'text'.}

\item{clean_field}{character. Default is 'text'. Name given to a new column in \code{df}
that will contain the cleaned text.}

\item{tolower}{logical. Lowercase text? Default is TRUE.}

\item{remove_punct}{logical. Remove punctuation?  Default TRUE.}

\item{simplify_punct}{logical. Replace all !, :, ?, ... by a single . Default FALSE.}

\item{replace_emojis}{logical. Replace a list of emojis by German words for certain
emotions ('wut', 'angst', 'freude'). Default TRUE.}

\item{replace_numbers}{logical. Replace numbers by German words for those numbers.
Default TRUE.}

\item{remove_stopwords}{logical. Remove German stopwords from the 'nltk' list with
the exception of some words defined by the author. Default FALSE.}

\item{store_uncleaned}{logical. Save uncleaned text as 'uncleaned_text' column?
Default TRUE.}

\item{count}{logical. Count number of words and strings after cleaning? Will create
the columns 'n_words' and 'n_chars'. Default TRUE.}
}
\value{
A data.frame with the same columns as the df. The cleaned text is in the
column specified by \code{text_field} or, if not specified in 'text.
}
\description{
Cleans text stored in a data frame. Function is tailored for German texts.
}
\examples{
tw_data \%<>\% head(10) \%>\% clean_text(text_field = 'full_text')
tw_data$text
}
