% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%<>\%}
\alias{\%<>\%}
\title{Assignment pipe}
\arguments{
\item{lhs}{An object which serves both as the initial value and as target.}

\item{rhs}{a function call using the magrittr semantics.}
}
\description{
See \code{magrittr::\link[magrittr:compound]{\%<>\%}} for details.
}
\keyword{internal}
