% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_similar_words.R
\name{detect_similar_words}
\alias{detect_similar_words}
\title{Detect similar words}
\usage{
detect_similar_words(
  word_df,
  model,
  word_field = "words",
  compare_by = NULL,
  compare_hits = T,
  min_simil = 0.7
)
}
\arguments{
\item{word_df}{A \link{data.frame} containing a column with words or multi-word expressions.}

\item{model}{A fastText model, loaded by \code{\link[fastrtext]{load_model}}.}

\item{word_field}{Character. The name of the column in word_df that contains the words.}

\item{compare_by}{Character. Default \code{NULL}. The name of a column that should be compared.}

\item{compare_hits}{Logical. Default \code{TRUE}. If true counts how often one word 'beats'
its similar other in regard to occurrences.}

\item{min_simil}{Numerical (0-1). Default .7. Similarity threshold. Word pairs below
this threshold are considered dissimilar.}
}
\value{
A data.frame. Containing a pairwise similarity table of all similar words.
}
\description{
Detects similar words in a data.frame, using a fastText model.
If requested, compares similar words along a variable specified in \code{compare_by},
and by the number of occurrences stored in the variable named \code{hits}.
Counts how often one word 'beats' its similar other in pairwise comparison.
This count is returned as 'wins_'.
}
\details{
Forces the 'word_df' to have a unique identifying variable called 'word_id';
re-uses any variable named 'id' that is unique.
}
\examples{
model <- fastrtext::load_model(system.file("extdata",
"tw_demo_model_sml.bin",
package = "dictvectoR"))
word_df <- data.frame(words = c("unsere steuern",
"steuerzahler",
 "unsere",
 "steuern"), hits = c(2, 3, 15, 4))
detect_similar_words(word_df, model)
}
