% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filter_ntile}
\alias{filter_ntile}
\title{Filter by ntile.}
\usage{
filter_ntile(df, var_field, probs)
}
\arguments{
\item{df}{A \link{data.frame}.}

\item{var_field}{A character string indicating the name of the variable used
for filtering.}

\item{probs}{A numerical value between 0 and 1 (i.e. proportion) indicating the
quantile threshold used for filtering.}
}
\value{
A \link{data.frame}, containing all rows for which the specified variable greater
or equal the specified threshold.
}
\description{
Filter a data frame for the top quantile of a specified variable.
}
\examples{
df <- data.frame(a = rnorm(10), b = sample(letters, 10, replace = TRUE))
filter_ntile(df, "a", .75)
}
