% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_distinctive.R
\name{find_distinctive}
\alias{find_distinctive}
\title{Find distinctive keywords}
\usage{
find_distinctive(
  df,
  concept_field,
  text_field = "text",
  word_df,
  word_field = "words",
  model
)
}
\arguments{
\item{df}{A data.frame containing one annotated document per row.}

\item{concept_field}{character. Name of the column that contains a binary,
(hand-coded) indicator of the presence of absence of a concept.}

\item{text_field}{character. Name of column in \code{df} that contains
the text of the documents. Default is "text".}

\item{word_df}{A \link{data.frame} containing a column with words or multi-word expressions.}

\item{word_field}{character. The name of the column in word_df that contains the words.}

\item{model}{A fastText model, loaded by \code{\link[fastrtext:load_model]{fastrtext::load_model()}}.}
}
\description{
Compares a word-vector representations of words to the representations of
an annotated data.frame of texts.
Aims to detect words that distinctively characterize a concept.
}
\details{
Takes an annotated data.frame \code{df} of texts as input.
The varialbe specified by \code{concept_field} in this df indicates the presence or
absence of a theoretical concept in the text.
Two average word-vector representations are computed, using a fastText \code{model}:
One for all texts that contain the concept, and one for those that do not.
A second data.frame, \code{word_df}, contains one (multi-)word per row in \code{'word_field'}.
Three new columns in \code{word_df} are created: The first, ending with \code{'_possim'},
indicates the cosine similarity between the
word and the positive concept corpus. The second \code{'_negsim'},
indicates the similarity the remaining corpus.
The third, ending with \code{'_distinctive'} is the difference between the two.
}
\examples{
model <- fastrtext::load_model(system.file("extdata",
                               "tw_demo_model_sml.bin",
                                package = "dictvectoR"))
tw_annot \%<>\% clean_text(text_field = "full_text")
word_df <- data.frame(words = c("skandal", "deutschland", "wundervoll"))
find_distinctive(tw_annot,
                 "pop",
                  word_df = word_df,
                  model = model)
}
\seealso{
\code{\link[=get_corpus_representation]{get_corpus_representation()}}, \code{\link[=find_distinctive]{find_distinctive()}}
}
