% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{find_unique_id}
\alias{find_unique_id}
\title{Find index of unique id in df.}
\usage{
find_unique_id(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
A numeric, indicating the index of the unique id column.
}
\description{
Searches a data.frame for a column that includes 'id' in its name and
is an unique identifier.
Returns the index of the first column in the data.frame that meets these conditions.
}
\examples{
df <- data.frame(id_not_unique = rep(1, 10),
unique_id = 1:10,
also_id = 21:30,
other = letters[1:10])
find_unique_id(df)

}
