% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_F1.R
\name{get_ARPF}
\alias{get_ARPF}
\title{Get Accuracy, Recall, Precision, and F1}
\usage{
get_ARPF(reference, prediction)
}
\arguments{
\item{reference}{A binary reference vector.}

\item{prediction}{A binary prediction vector.}
}
\value{
A data.frame with columns 'accuracy', 'recall', 'precision', and 'F1'.
}
\description{
Returns Accuracy, Recall, Precision, and F1 scores for a binary reference variable
and a binary prediction variable.
}
\details{
The values of \code{reference} and \code{prediction} must be 0 or 1.
Both variables can be stored in a data.frame or as vectors of equal length.
Can be of type numeric or factor.
}
\examples{
mtcars$pred <- get_prediction(mtcars, mtcars$am, mtcars$drat)
get_ARPF(mtcars$am, mtcars$pred)
}
\seealso{
\code{\link[=confuse]{confuse()}}, \code{\link[=get_prediction]{get_prediction()}}, \code{\link[caret:confusionMatrix]{caret::confusionMatrix()}}
}
