% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combinations.R
\name{get_combis}
\alias{get_combis}
\title{Get combinations of keywords}
\usage{
get_combis(
  word_df,
  word_field = "words",
  dims = NULL,
  min_per_dim = 1,
  max_overall,
  limit = NULL,
  seed = 1,
  save_settings = T,
  save_input = F
)
}
\arguments{
\item{word_df}{A \link{data.frame} containing a column with words or multi-word expressions.}

\item{word_field}{character. Default is \code{'words'}. The name of the column in
word_df that contains the words.}

\item{dims}{character. Default is \code{NULL}. The name of the column in word_df that groups
the words into dimensions. Can be ignored.}

\item{min_per_dim}{numerical. Default is \code{1}. Minimum number of words (per dimension)
returned. Will be replaced by 1 if < 1.}

\item{max_overall}{numerical. Maximum number of words per returned combination.}

\item{limit}{numerical. Default is \code{NULL}. Limits the number of combinations of equal
length per dimension by randomization.}

\item{seed}{numerical. Default is \code{1}. Input to make randomization reproducible.}

\item{save_settings}{logical. Default is \code{TRUE}. Saves randomization settings in
df to make it reproducible.}

\item{save_input}{logical. Default is \code{FALSE}. Saves input words and dims in df as
character string.}
}
\value{
A data.frame of combinations.
The combinations are stored as a \link{list} of character vectors in \code{combs_split}.
Use this column to pass it to \code{\link{get_many_F1s}} or \code{\link{get_many_RPFs}}.

The data.frame will include string variables for the words in the combinations
called \code{combs} for each dimension.
Do not use these columns passing the dictionaries on to \code{\link{get_many_F1s}},
as this will result in a
faulty average representation, caused by how representations for multi-word expressions
are queried.

Additionally, the data.frame includes a rowid, a count variable for the number of words
overall (\code{sum_nterms}),
and counts for each dimension, which can be used to remove imbalanced dictionaries.
If requested, \code{settings} stores the randomization settings,
and \code{input} the words and dimensions used as input.
}
\description{
Returns combinations of keywords.
}
\details{
Takes a data.frame \code{word_df} with a character column specified by \code{words}
as input.
As default, it will return all combinations of various lengths of these words.
Additionally, the function can account for conceptual dimensions, identified by
a categorical (character, numerical, or factor)
column in \code{word_df} specified by \code{dims}. If dimensions are specified, the function
will find all combinations for each dimension, and
will return all combinations of these combinations. CAUTION: This can lead quickly
to an extremely large number of combinations.

The number of combinations can be limited in several ways:

Firstly, the minimum number of words returned per dimension is specified by \code{min_per_dim}, the maximum number of words overall is set by \code{max_overall}.

Secondly, the function allows for random sampling of combinations.
This is recommended, as it drastically reduces the number of returned combinations,
reduces the computational load, and speeds up the process. (Of course, this comes at
the cost of completeness).
Random sampling is implemented using the \code{\link[RcppAlgos]{comboSample}} function.
Setting a \code{limit} will cap the number of combinations
of equal length for each dimension.

E.g., if \code{limit = 5}, \code{min_per_dim = 2}, \code{max_overall = 6} is set for a \code{word_df}
containing two dimensions \code{a} and \code{b},
the function will pick max. five combinations of length 2 for \code{a}, and five of
length 2 for \code{b}, five of length 3 for \code{a},
and five of length 3 for \code{b}, and will return all combinations of these combinations.
}
\examples{
test_df <- data.frame(words = letters[1:8],
                      dim = rep(paste0("c_", 1:2), 4))
t0 <- get_combis(test_df,
                 word_field = "words",
                 dims = "dim",
                 max_overall = 5)
t1 <- get_combis(test_df,
                 word_field = "words",
                 dims = "dim",
                 max_overall = 5,
                 limit = 5)
}
