% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_distinctive.R
\name{get_corpus_representation}
\alias{get_corpus_representation}
\title{Vector representation of a corpus.}
\usage{
get_corpus_representation(df, model, text_field = "text", normalize = T)
}
\arguments{
\item{df}{A \link{data.frame} with a column containing text identified by \code{text_field}.}

\item{model}{A fastText model, loaded by \code{\link[fastrtext:load_model]{fastrtext::load_model()}}.}

\item{text_field}{A \link{character} string indicating the name of the text column in df.}

\item{normalize}{Logical. Default \code{TRUE}. Normalize the vectors to their Euclidean norm?}
}
\value{
A single-row sparse matrix of class \linkS4class{dgCMatrix}
as returned by \code{\link[Matrix]{Matrix}}.
}
\description{
Returns the average word-vector representation of a text column in a data frame,
using a \code{fastText} model.
}
\examples{
model <- fastrtext::load_model(system.file("extdata",
                                           "tw_demo_model_sml.bin",
                                            package = "dictvectoR"))
tw_annot <- tw_annot \%>\% head(15) \%>\% clean_text(text_field = "full_text")
corpus_rep <- get_corpus_representation(tw_annot, model)
}
\seealso{
\code{\link[fastrtext]{get_sentence_representation}}, \code{\link{get_word_representations}}
}
