% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_F1.R
\name{get_F1}
\alias{get_F1}
\title{Get F1 score for a DDR measure}
\usage{
get_F1(
  df,
  dictionary,
  model,
  reference,
  text_field = "text",
  replace_na = c("mean-sd", "min", 0, F)
)
}
\arguments{
\item{df}{A data.frame containing one annotated document or sentence per row.}

\item{dictionary}{A character vector containing the keywords of a dictionary,
passed to \code{\link{cossim2dict}}.}

\item{model}{A fasttext model as loaded by \code{\link[fastrtext]{load_model}}.}

\item{reference}{Name of the binary reference column in df.}

\item{text_field}{Name of column in \code{df} that contains the text of the documents.
Default is "text".}

\item{replace_na}{Specifies the value used to replace NAs in the DDR measurement.
Default is 'mean-sd'. Can take values:
\itemize{
\item \code{'mean-sd'} (charcter): replace NAs by mean - 1sd. Default.
\item \code{'min'} (charcter): replace NAs by minimum.
\item \code{0} (numerical): replace NAs by 0.
\item \code{FALSE} (logical): do not replace NAs.
}}
}
\description{
Returns the F1 score for a DDR measurement in predicting a binary reference
(i.e. a manually annotated variable).
}
\details{
The gradual DDR measurement is passed to \code{\link[=get_prediction]{get_prediction()}} to obtain a binary
prediction through logistic regression.
The F1 scores indicate the performance of these words/dictionaries in predicting a binary
coding, when used in the DDR method.
The F1 score is the harmonic mean between Recall and Precision (1).
}
\examples{
model <- fastrtext::load_model(system.file("extdata",
                               "tw_demo_model_sml.bin",
                                package = "dictvectoR"))
tw_annot \%<>\% clean_text(text_field = "full_text")
dict <- c("skandal", "deutschland", "steuerzahler")
get_F1(tw_annot, dict, model, 'pop')
}
\references{
(1) Chinchor, N. (1992). MUC-4 evaluation metrics. Proceedings of the 4th
Conference on Message Understanding, 22–29.
\url{https://doi.org/10.3115/1072064.1072067}
}
\seealso{
\code{\link[=cossim2dict]{cossim2dict()}}, \code{\link[=get_prediction]{get_prediction()}}, \code{\link[=get_many_F1s]{get_many_F1s()}}, \code{\link[=get_many_RPFs]{get_many_RPFs()}}, \code{\link[caret:confusionMatrix]{caret::confusionMatrix()}}
}
