% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mulitiwords.R
\name{get_hits}
\alias{get_hits}
\title{Get occurrence frequency of words.}
\usage{
get_hits(word_df, tokens, word_field = "words", replace = F)
}
\arguments{
\item{word_df}{A data.frame containing words.}

\item{tokens}{A word-tokens object, returned by quanteda::tokens.}

\item{word_field}{Character. Default is "words". Name of the column in
\code{word_df} that contains the words.}

\item{replace}{Logical. Default is \code{FALSE}. If \code{FALSE} checks fills in 'hits'
only for missing observations. If \code{TRUE} counts 'hits' for all words in word_df.}
}
\value{
A data.frame with a column \code{'hits'} indicating the frequency of this term
in the tokens. Arranged by descending number of hits.
}
\description{
Adds the number of occurrences of a word or multi-word expression
in a quanteda tokens object to a data.frame.
By default, it checks if 'hits' have been counted before and only fills
in missing values. Works with GLOB-style wildcards.
}
\examples{
tw_data \%<>\% head(100) \%>\% clean_text(text_field = 'full_text')
toks <- quanteda::tokens(tw_data$text)
word_df <- data.frame(words = c("der deutschen", "steuer*", "xyz")) \%>\%
get_hits(tokens = toks)
}
