% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_F1.R
\name{get_many_F1s}
\alias{get_many_F1s}
\title{Get F1 scores for many words or dictionaries.}
\usage{
get_many_F1s(
  words,
  model,
  df,
  reference,
  text_field = "text",
  replace_na = c("mean-sd", "min", 0, F)
)
}
\arguments{
\item{words}{A character vector containing keywords, or a list of character vectors
containing dictionaries.}

\item{model}{A fastText model, loaded by \code{\link[fastrtext:load_model]{fastrtext::load_model()}}.}

\item{df}{A data.frame containing one annotated document per row.}

\item{reference}{Name of the binary reference column in \code{df} (character).}

\item{text_field}{Name of column in \code{df} that contains the text of the documents.
Default is "text".}

\item{replace_na}{Specifies the value used to replace NAs in the DDR measurement.
Default is 'mean-sd'. Can take values:
\itemize{
\item \code{'mean-sd'} (charcter): replace NAs by mean - 1sd. Default.
\item \code{'min'} (charcter): replace NAs by minimum.
\item \code{0} (numerical): replace NAs by 0.
\item \code{FALSE} (logical): do not replace NAs.
}}
}
\description{
Efficiently computes F1 scores for all elements of a vector containing keywords,
or a list containing dictionaries, when used in DDR method.
}
\details{
A numerical F1 score is returned for each element (i.e. word or dictionary) of
the vector or list.
The F1 scores indicate the performance of these words/dictionaries in predicting
a binary coding, when used in the DDR method.
The resulting gradual measure from the DDR measure is passed to a logistic regression,
with the binary coding as dependent variable.
Binary predictions are calculated from this logistic model and compared with the binary
coding.
The F1 score is the harmonic mean between Recall and Precision (1).
}
\examples{
model <- fastrtext::load_model(system.file("extdata",
                               "tw_demo_model_sml.bin",
                                package = "dictvectoR"))
tw_annot \%<>\% clean_text(text_field = "full_text")
dict_df <- data.frame(id = 1:3)
dict_df$combis <- list(c("mehrheit deutschen", "merkel", "skandal"),
                      c("steuerzahler", "bundesregierung",
                      "komplett gescheitert"),
                      c( "arbeitnehmer", "groko", "wahnsinn"))
dict_df$F1 <- get_many_F1s(dict_df$combis,
                           model = model,
                           df = tw_annot,
                           reference = "pop")
}
\references{
(1) Chinchor, N. (1992). MUC-4 evaluation metrics. Proceedings of the 4th
Conference on Message Understanding, 22–29.
\url{https://doi.org/10.3115/1072064.1072067}
}
\seealso{
\code{\link[=cossim2dict]{cossim2dict()}}, \code{\link[=get_prediction]{get_prediction()}}, \code{\link[=get_F1]{get_F1()}}, \code{\link[=get_many_RPFs]{get_many_RPFs()}}
}
