% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_F1.R
\name{get_many_F1s_by_group}
\alias{get_many_F1s_by_group}
\title{Get F1 for many by a grouping varialbe}
\usage{
get_many_F1s_by_group(
  keyword_df,
  keyword_field = "words",
  id = "id",
  model,
  text_df,
  group_field,
  reference,
  text_field = "text",
  replace_na = c("mean-sd", "min", 0, F)
)
}
\arguments{
\item{keyword_df}{A data.frame, containing a column with a character vector of words,
or with a list of dictionaries.}

\item{keyword_field}{character. The name of the column in \code{keyword_df} that is either
a character vector of single keywords,
or a \link{list} of dictionaries, stored as separate character vectors with one element
per word.}

\item{id}{A unique identifier for each keyword.}

\item{model}{A fastText model, loaded by \code{\link[fastrtext]{load_model}}.}

\item{text_df}{A data.frame containing one annotated document per row.}

\item{group_field}{character. Name of a categorical grouping variable in \code{df}.}

\item{reference}{character. Name of the binary reference column in \code{df}.}

\item{text_field}{Name of column in \code{df} that contains the text of the documents.
Default is "text".}

\item{replace_na}{Specifies the value used to replace NAs in the DDR measurement.
Default is 'mean-sd'. Can take values:
\itemize{
\item \code{'mean-sd'} (charcter): replace NAs by mean - 1sd. Default.
\item \code{'min'} (charcter): replace NAs by minimum.
\item \code{0} (numerical): replace NAs by 0.
\item \code{FALSE} (logical): do not replace NAs.
}}
}
\description{
Efficiently computes F1 scores for a character vector of keywords stored in a
data.frame,
or a \link{list} of dictionaries stored in a data.frame - for a reference data.frame
grouped by \code{group_field}.
ADD DETAILS....
}
\details{
#'@seealso \code{\link{cossim2dict}}, \code{\link{get_prediction}}, \code{\link{get_F1}}, \code{\link{get_many_RPFs}}, \code{\link[caret]{confusionMatrix}}
}
\examples{
model <- fastrtext::load_model(system.file("extdata",
                               "tw_demo_model_sml.bin",
                                package = "dictvectoR"))
tw_annot \%<>\% clean_text(text_field = "full_text")
dict_df <- data.frame(id = 1:3)
dict_df$combis <- list(c("mehrheit deutschen", "merkel", "skandal"),
                      c("steuerzahler", "bundesregierung",
                      "komplett gescheitert"),
                      c( "arbeitnehmer", "groko", "wahnsinn"))
get_many_F1s_by_group(keyword_df = dict_df,
                     keyword_field = "combis",
                     id = "id",
                     model = model,
                     text_df = tw_annot,
                     group_field = "party",
                     reference = 'pop')

}
\seealso{
\code{\link{cossim2dict}}, \code{\link{get_prediction}}, \code{\link{get_many_F1s}}
}
