% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_F1.R
\name{get_many_RPFs}
\alias{get_many_RPFs}
\title{Get Recall, Precision, F1 for many}
\usage{
get_many_RPFs(
  keyword_df,
  keyword_field = "words",
  model,
  text_df,
  reference,
  text_field = "text",
  replace_na = c("mean-sd", "min", 0, F)
)
}
\arguments{
\item{keyword_df}{A data.frame, containing a column with a
character vector of words, or with a list of dictionaries.}

\item{keyword_field}{character. The name of the column in \code{keyword_df} that is
either a character vector of single keywords,
or a \link{list} of dictionaries, stored as separate character vectors with one
element per word.}

\item{model}{A fastText model, loaded by \code{\link[fastrtext]{load_model}}.}

\item{text_df}{A data.frame containing one annotated document per row.}

\item{reference}{Name of the binary reference column in \code{df} (character).}

\item{text_field}{Name of column in \code{df} that contains the text of the documents.
Default is "text".}

\item{replace_na}{Specifies the value used to replace NAs in the DDR measurement.
Default is 'mean-sd'. Can take values:
\itemize{
\item \code{'mean-sd'} (charcter): replace NAs by mean - 1sd. Default.
\item \code{'min'} (charcter): replace NAs by minimum.
\item \code{0} (numerical): replace NAs by 0.
\item \code{FALSE} (logical): do not replace NAs.
}}
}
\description{
Efficiently computes Recall, Precision, and F1 scores for a character vector of
keywords stored in a data.frame,
or a \link{list} of dictionaries stored in a data.frame. Adds Recall, Precision, and F1
to the data.frame.
}
\details{
For each element (i.e. word or dictionary) of the character vector or list
in the data.frame .
The F1 scores indicate the performance of these words/dictionaries in predicting a
binary coding, when used in the DDR method.
The resulting gradual measure from the DDR measure is passed to a logistic regression,
with the binary coding as dependent variable.
Binary predictions are calculated from this logistic model and compared with the
binary coding.
The F1 score is the harmonic mean between Recall and Precision (Chinchor, 1992).
}
\examples{
model <- fastrtext::load_model(system.file("extdata",
                               "tw_demo_model_sml.bin",
                                package = "dictvectoR"))
tw_annot \%<>\% clean_text(text_field = "full_text")
dict_df <- data.frame(id = 1:3)
dict_df$combis <- list(c("mehrheit deutschen", "merkel", "skandal"),
                      c("steuerzahler", "bundesregierung",
                      "komplett gescheitert"),
                      c( "arbeitnehmer", "groko", "wahnsinn"))
get_many_RPFs(keyword_df = dict_df,
       keyword_field = "combis",
       model = model,
       text_df = tw_annot, reference = "pop")
}
\references{
Chinchor, N. (1992). MUC-4 evaluation metrics. Proceedings of the
4th Conference on Message Understanding, 22–29.
\url{https://doi.org/10.3115/1072064.1072067}
}
\seealso{
\code{\link{cossim2dict}}, \code{\link{get_prediction}}, \code{\link{get_F1}}, \code{\link{get_many_RPFs}}, \code{\link[caret]{confusionMatrix}}

\code{\link{cossim2dict}}, \code{\link{get_prediction}}, \code{\link{get_many_F1s}}
}
