% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_F1.R
\name{get_prediction}
\alias{get_prediction}
\title{Get binary prediction}
\usage{
get_prediction(data, dv, iv)
}
\arguments{
\item{data}{A data frame.}

\item{dv}{A binary variable in that data frame.}

\item{iv}{A gradual variable in that data frame used for predicting \code{dv}.}
}
\value{
A factor with levels 0, 1.
}
\description{
Returns the prediction of a binary variable by a gradual measurement using
logistic regression.
}
\details{
Predictions are considered '1' if the predicted probability is >= .5.
}
\examples{
data(mtcars)
mtcars$pred <- get_prediction(mtcars, mtcars$am, mtcars$drat)
}
\seealso{
\code{\link[stats:glm]{stats::glm()}}
}
