% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_distinctive.R
\name{get_word_representations}
\alias{get_word_representations}
\title{Get word representations.}
\usage{
get_word_representations(word_df, model, word_field = "words", normalize = T)
}
\arguments{
\item{word_df}{A \link{data.frame} containing a column with words or multiword expressions.}

\item{model}{A fastText model, loaded by \code{\link[fastrtext:load_model]{fastrtext::load_model()}}.}

\item{word_field}{A \link{character} string indicating the name of the column
in word_df that contains the words.}

\item{normalize}{Logical. Default \code{TRUE}. Normalize the vectors to their Euclidean norm?}
}
\value{
A sparse matrix of class \linkS4class{dgCMatrix} as
returned by \code{\link[Matrix]{Matrix}}.
Has the same number of rows as \code{word_df} and same number of columns
as dimensions of the \code{model}.
}
\description{
Wrapper around \code{\link[fastrtext]{get_sentence_representation}} to return
the \code{fastText} word-vector representation
of words or multiwords, stored in a column of a data frame \verb{'word_df}.
}
\examples{
model <- fastrtext::load_model(system.file("extdata",
                                           "tw_demo_model_sml.bin",
                                            package = "dictvectoR"))
word_df <- data.frame(words = c("das ist", "ein", "test"))
word_rep <- get_word_representations(word_df, model)
}
\seealso{
\code{\link[fastrtext]{get_sentence_representation}}, \code{\link{get_corpus_representation}}
}
