% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector}
\usage{
normalize(x)
}
\arguments{
\item{x}{a numeric vector, or a matrix of vectors.}
}
\description{
Normalizes a (matrix) vector to its Euclidean norm, or 'L2'-norm.
}
\details{
The norm is defined as
\deqn{||x|| = \sqrt{\sum(x^2)}}
The Euclidean norm can be interpreted as the length of the connection
between zero and a point specified by the vector.
The normalized vector, then, is computed as:
\deqn{x' = x/||x||}
}
\examples{
v <- rnorm(10)
normalize(v) == v/sqrt(sum(v^2))
}
