% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{prepare_train_data}
\alias{prepare_train_data}
\title{Prepare text for fastText-model-training}
\usage{
prepare_train_data(df, text_field = "text", seed = 1)
}
\arguments{
\item{df}{A data frame.}

\item{text_field}{Name of a column in that \code{df} that contains text that should be
cleaned. Default is 'text'.}

\item{seed}{Used for random shuffling. Default is 1.}
}
\value{
A character vector that can be directly used to train a fasttext model
with \code{\link[fastrtext]{build_vectors}}.
}
\description{
Helper function to prepare text for training a fastText model.
(Caution: Tailored to texts in German language).
}
\details{
Takes a data.frame containing text. First checks the length of
the texts specified in \code{text_field} using \code{\link[quanteda]{nsentence}}.
Texts with more than 3 sentences are tokenized by \code{\link[quanteda]{tokens}}
into sentences.
All texts are passed to \code{\link{clean_text}} with the fixed settings:
\itemize{
\item \code{tolower = T}
\item \code{remove_punct = T}
\item \code{replace_emojis = T}
\item \code{replace_numbers = T}
\item \code{remove_stopwords = F}
\item \code{store_uncleaned = F}
\item \code{count = T}
}
The cleaned, short texts are shuffled and returned as a character vector.
}
\examples{
texts <- prepare_train_data(head(tw_data, 10), text_field = 'full_text')

}
\seealso{
\code{\link[fastrtext]{build_vectors}}, \code{\link[quanteda]{tokens}}, \code{\link{clean_text}}
}
