% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_similar_words.R
\name{remove_similar_words}
\alias{remove_similar_words}
\title{Remove too similar terms}
\usage{
remove_similar_words(
  word_df,
  model,
  word_field = "words",
  compare_by = NULL,
  compare_hits = T,
  min_simil = 0.7,
  win_threshold = 0.5
)
}
\arguments{
\item{word_df}{A \link{data.frame} containing a column with words or multi-word expressions.}

\item{model}{A fastText model, loaded by \code{\link[fastrtext]{load_model}}.}

\item{word_field}{character. The name of the column in word_df that contains the words.}

\item{compare_by}{character. Default \code{NULL}. The name of a column that should be compared.}

\item{compare_hits}{logical. Default \code{TRUE}. If true counts how often one word 'beats'
its similar other in regard to occurrences.}

\item{min_simil}{Numerical (0-1). Default .7. Similarity threshold. Word pairs below this
threshold are considered dissimilar.}

\item{win_threshold}{Numerical (0-1). Default .5. Determines the threshold to drop words,
defined as proportion of won pairwise comparisons;
resp., if both \code{compare_by} and \code{compare_hit} are set, the mean of both proportional wins.}
}
\value{
A data.frame. Containing a pairwise similarity table of all similar words.
}
\description{
Removes similar terms from a data.frame \code{word_df}.
}
\details{
Detects similar words and multiword expressions in \code{word_df}, using a \code{fastText} model.
The cosine similarity threshold is set by \code{min_simil}.
If requested, similar words are compared along a variable specified in \code{compare_by},
and/or by the number of occurrences stored in the variable named \code{hits}.

The threshold for dropping terms is set by \code{win_threshold}.
This indicates the proportion of how many of the pairwise comparisons are 'won'
by the word in question. If \code{compare_hits = T} and a \code{compare_by} is set,
it is the mean of both proportions. If a word 'wins' all comparisions
regarding frequency (wins_hits1 == 1),
but loses all comparisions regaring the set score ('wins_score1' == 0), this value is .5.

Forces \code{'word_df'} to contain a unique identifier called 'word_id';
re-uses the first variable named 'id' that is such a unique identifier.
}
\examples{
model <- fastrtext::load_model(system.file("extdata",
                                          "tw_demo_model_sml.bin",
                                          package = "dictvectoR"))
set.seed(1)
word_df <- data.frame(words = c("unsere steuern",
                               "steuerzahler",
                               "unsere",
                               "steuern"),
                     hits = c(2, 3, 15, 4),
                     score = rnorm(4))
remove_similar_words(word_df,
                    model,
                    compare_by = "score",
                    compare_hits = FALSE,
                    win_threshold = .4)
}
