% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{repl_na}
\alias{repl_na}
\title{Replace missing values.}
\usage{
repl_na(x, replace_na = c("mean-sd", "min", 0, F))
}
\arguments{
\item{x}{A numerical vector.}

\item{replace_na}{Specifies the value used to replace NAs. Can take values:
\itemize{
\item 'mean-sd' (charcter): replace NAs by mean - 1sd. Default.
\item 'min' (charcter): replace NAs by minimum.
\item 0 (numerical): replace NAs by 0.
\item FALSE (logical): do not replace NAs.
}}
}
\description{
Helper to replace missing values. Default is to repalce NAs by 'mean' - 1 'sd'.
}
\examples{
a <- c(rnorm(7), NA, NA, NA)
repl_na(a) == repl_na(a, 'mean-sd')
repl_na(a, 'min')
repl_na(a, 0)
repl_na(a, FALSE)
}
