% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_distinctive.R
\name{simil_words2rep}
\alias{simil_words2rep}
\title{Cosine similarity between words and a given vector.}
\usage{
simil_words2rep(word_df, word_field = "words", rep, model)
}
\arguments{
\item{word_df}{A \link{data.frame} containing a column with words or multiword expressions.}

\item{word_field}{A \link{character} string indicating the name of the column in
word_df that contains the words.}

\item{rep}{A given word-vector representation, stored as numerical vector,
matrix, or sparse matrix object. \code{length}, resp. \code{ncol} of \code{rep} must be equal
to the dimensions of the used \code{fastText} model.}

\item{model}{A fastText model, loaded by \code{\link[fastrtext:load_model]{fastrtext::load_model()}}.}
}
\description{
Returns the cosine similarity of each word in a data frame and a given vector
representation.
}
\examples{
model <- fastrtext::load_model(system.file("extdata",
                               "tw_demo_model_sml.bin",
                                package = "dictvectoR"))
pop_rep <- tw_annot \%>\%
           dplyr::filter(pop == 1) \%>\%
           clean_text(text_field = "full_text") \%>\%
           get_corpus_representation(model = model)
words_df <- data.frame(words = c("coronadeutschland", "skandal"))
words_df$popsimil <- simil_words2rep(words_df,
                                     word_field = "words",
                                     rep = pop_rep,
                                     model)
}
\seealso{
\code{\link{get_corpus_representation}}, \code{\link[=find_distinctive]{find_distinctive()}}
}
