% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tw_data}
\alias{tw_data}
\title{Tweets from German politicians.}
\format{
A data frame with 20838 rows and 8 variables:
\describe{
\item{user_id}{Twitter user ID}
\item{twitter_handle}{Twitter handle}
\item{party}{Political party}
\item{followers_count}{The number of followers in Aug 2022}
\item{status_id}{Tweet ID}
\item{created_at}{Date and time when the Tweet was created}
\item{full_text}{Uncleaned text of the Tweet}
\item{poppa_populism}{Mean populism score from POPPA expert survey}
}
}
\source{
\href{https://developer.twitter.com/}{Twitter},
\href{https://doi.org/10.7802/2415}{EPINet Twitter Dataset},
\href{https://doi.org/10.7910/DVN/8NEL7B}{POPPA}
}
\usage{
tw_data
}
\description{
A dataset containing 20,838 Tweets from 32 German politicians,
posted between 2020-03-11 and 2021-09-25.
}
\details{
The Twitter accounts were selected from the EPINetz Twitter Politicians Dataset 2021 (König et al., 2022).
For each of the seven political parties represented in the German Parliament, the five most popular
Twitter accounts from politicians active at the federal level or party representatives were selected.
Only accounts with more than 30,000 followers were selected.

The timeframe starts with the beginning of the Covid-19 pandemic and ends one day before
the German general elections 2021.

For each account, the maximum number (3,200) of Tweets returned by the API v2 were downloaded in August 2022,
using rtweet (Kearney, 2022). Quotes, re-tweets, and Tweets outside the timeframe were excluded.
Completeness of this dataset cannot be guaranteed.

The dataset also includes a variable extracted from the POPPA Populism and Political Parties Expert Survey,
indicating the mean expert rating of populism per political party.
}
\references{
Kearney, M. W., Sancho, L. R., Wickham, H., Heiss, A., Briatte, F., & Sidi, J. (2022).
rtweet: Collecting Twitter Data. Retrieved from \url{https://CRAN.R-project.org/package=rtweet}

König, T., Schünemann, W. J., Brand, A., Freyberg, J., & Gertz, M. (2022).
The EPINetz Twitter Politicians Dataset 2021. A New Resource for the Study
of the German Twittersphere and Its Application for the 2021 Federal Elections.
Politische Vierteljahresschrift. \url{https://doi.org/10.1007/s11615-022-00405-7}

Meijers, M., & Zaslove, A. (2020). Populism and Political Parties Expert Survey 2018 (POPPA) (Data set).
Harvard Dataverse. \url{https://doi.org/10.7910/DVN/8NEL7B}
}
\keyword{datasets}
